Rename psi widget to PageSpeed Insights and update folder structure
