+++ components/widgets/pagespeed-insights/score.js
